+++ views/projects/edit.ejs
@@ -1,0 +1,27 @@
+<%- include('../partials/header') %>
+
+<form id="project-edit-form" action="/projects/<%=project._id%>?_method=PUT" method="POST">
+
+    <label>Title:</label>
+  <input type="text" name="title" value="<%=project.title%>">
+
+  <label>Description:</label>
+  <textarea name="description"><%=project.description%></textarea>
+
+  <label>Due:</label>
+  <input type="datetime-local" name="due" value="<%=project.due%>">
+
+  <label>Priority</label>
+  <select name="priority">
+    <option value="P1">P1</option>
+    <option value="P2">P2</option>
+    <option value="P3">P3</option>
+    <option value="P4">P4</option>
+    <option value="P5">P5</option>
+  </select>
+
+  <!-- <label>Invite Team Members:</label>
+  <input type="email" name="invitation-email-address"> -->
+
+  <input type="submit" value="Save Changes">
+</form>
