+++ views/projects/show.ejs
@@ -1,7 +1,16 @@
 <%- include('../partials/header') %>
 
 <section id="show-page">
-  <h1>Show Page</h1>
+  <h1><%= project.title %></h1>
+  <p><%= project.description %></p>
+  <div><%= project.owner %></div>
+  <div><%= project.priority %></div>
+  
+  <a href="/projects/<%=project._id%>/edit">Click to edit me</a>
+  <form id="project-delete-form" action="/projects/<%=project._id%>?_method=DELETE" method="POST">
+    <input type="submit" value="Delete Project">
+  </form>
+  
 </section>
   
 <%- include('../partials/footer') %>
